removed entity, primary key, jpa repository added custom repository with jdbc template

+++ src/main/kotlin/eu/wojciechzurek/pwnedpasswordchecker/PwnedPasswordResponse.kt
@@ -2,7 +2,7 @@ package eu.wojciechzurek.pwnedpasswordchecker
 
 data class PwnedPasswordResponse (
         val suffix: String,
-        val count: Long
+        val count: Int
 ){
     companion object {
         fun from(pwnedPassword: PwnedPassword) = PwnedPasswordResponse(pwnedPassword.suffix, pwnedPassword.count)
